[TM-1144] reorder functions to get valid schema data for polygons
